Testing/requirements.txt
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0  # If you're testing FastAPI endpoints 
fastapi>=0.109.0
sqlalchemy>=2.0.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
Pillow>=10.0.0  # For image processing 